+++ 02_Development/Display/sensors.py
@@ -2,7 +2,7 @@ import os
 import random
 from abc import ABC, abstractmethod
 
-from process_sample_data import ProcessSampleData
+from processsampledata import ProcessSampleData
 from rpi_check import is_on_raspberry_pi
 import constants
