--- Display/features/environment.py
@@ -1,8 +1,0 @@
-
-
-def after_step(context, step):
-    if step.status == "failed":
-        try:
-            context.client.containers.get(context.container_name).kill()
-        except NameError:
-            pass
