--- Display/runtests.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-
-echo "Behaviour Tests:"
-behave
-echo $'\nUnit Tests:'
-python3 -m unittest
